action_creators: move saveAndProcessPage and helpers to page.ts
